requirements.txt
snowflake-connector-python==4.3.0
python-dotenv==1.2.1
